requirements.txt
streamlit==1.28.1
firecrawl-py==0.0.16
python-dotenv==1.0.0
